refactor: window.py 平台层剥离（批6-3）
- Win32 平台细节（ctypes 结构体/点击穿透/PerPixelInputController/
  全屏探测）→ platform_win.py
- macOS 原生窗口辅助 → platform_mac.py
- window.py 净减 223 行；ctypes 声明逐字符保真；
  显式 re-export 维持全部调用点与测试桩兼容
- 5.6sol 盲审：P1（sys.platform monkeypatch seam）经实测证伪；
  P2（子类覆写 seam）记录为接受风险（仓库无 PetWindow 子类）

结构优化批6-3（STRUCT_PLAN_FINAL.md）
