+++ pet/window.py
@@ -14,7 +14,6 @@
 from __future__ import annotations
 
 import ctypes
-from ctypes import wintypes
 import json
 import logging
 import os
@@ -73,6 +72,25 @@ from .click_sound import (
 from .proactive import effective_proactive_config
 from .updater import QUARK_PAN_URL, REPO_URL
 
+from . import platform_win
+from .platform_mac import _keep_macos_tool_window_visible, _mac_set_window_level
+from .platform_win import (
+    GWL_STYLE as GWL_STYLE,
+    GWL_EXSTYLE as GWL_EXSTYLE,
+    _WS_CAPTION as _WS_CAPTION,
+    _WS_EX_TOPMOST as _WS_EX_TOPMOST,
+    _WS_EX_TRANSPARENT as _WS_EX_TRANSPARENT,
+    _WinRect as _WinRect,
+    _WinMonitorInfo as _WinMonitorInfo,
+    _set_windows_click_through as _set_windows_click_through,
+    WindowsPerPixelInputController as WindowsPerPixelInputController,
+    _FS_SKIP_CLASSES as _FS_SKIP_CLASSES,
+    _fullscreen_geometry_hit as _fullscreen_geometry_hit,
+    _fs_user_busy_state as _fs_user_busy_state,
+    _fg_fullscreen_probe as _fg_fullscreen_probe,
+    _fg_fullscreen_win32 as _fg_fullscreen_win32,
+)
+
 # 后台播放音乐时自动播放的唱歌/哼歌动画
 SING_ANIM = '悠闲哼歌'
 
@@ -97,74 +115,6 @@ def _resolve_self_talk_image_dir(raw: str) -> str:
     if candidate.is_absolute() and not candidate.is_dir():
         return ''
     return str(resolve_fun_asset(raw, oijingjing_image_path().parent))
-
-
-def _keep_macos_tool_window_visible(window) -> None:
-    """Tool windows must remain visible while another application is active.
-
-    This is independent from the configurable z-order. Without the attribute,
-    Cocoa automatically hides a Qt.Tool window when the accessory application
-    resigns active, which looked like the WebM Chat pet had exited.
-    """
-    if sys.platform == 'darwin':
-        window.setAttribute(Qt.WidgetAttribute.WA_MacAlwaysShowToolWindow, True)
-
-
-def _mac_set_window_level(view_id: int, level: int) -> bool:
-    """macOS 原生：把 NSWindow 层级设为指定值（3=置顶浮动，0=普通）。
-
-    Qt 的 WindowStaysOnTopHint 在 macOS 上对无边框 Tool 窗口/运行时切换不可靠，
-    这里用 objc runtime 直接调 [NSWindow setLevel:] 强制生效（ctypes 零依赖）。
-
-    只在真实 cocoa 平台执行：offscreen/minimal 等测试平台下 winId() 不是
-    NSView 指针，objc_msgSend 会直接 SIGSEGV（无法被 try/except 捕获）。
-    """
-    if sys.platform != 'darwin':
-        return False
-    try:
-        from PySide6.QtGui import QGuiApplication
-        if QGuiApplication.platformName() != 'cocoa':
-            return False
-    except Exception:
-        return False
-    try:
-        import ctypes
-        import ctypes.util
-
-        lib_path = ctypes.util.find_library('objc') or '/usr/lib/libobjc.A.dylib'
-        objc = ctypes.cdll.LoadLibrary(lib_path)
-
-        # 关键：sel_registerName 返回 SEL（64 位指针）。ctypes 默认按 c_int(32 位)
-        # 截断返回值，损坏的 SEL 会让 ObjC runtime 段错误（SIGSEGV），必须显式声明
-        objc.sel_registerName.restype = ctypes.c_void_p
-        objc.sel_registerName.argtypes = [ctypes.c_char_p]
-
-        msg = objc.objc_msgSend
-        msg.restype = ctypes.c_void_p
-
-        sel_window = objc.sel_registerName(b'window')
-        sel_set_level = objc.sel_registerName(b'setLevel:')
-        sel_order_front = objc.sel_registerName(b'orderFrontRegardless')
-
-        # [view window] —— 无参，返回 NSWindow*
-        msg.argtypes = [ctypes.c_void_p, ctypes.c_void_p]
-        window = msg(ctypes.c_void_p(view_id), sel_window)
-        if not window:
-            return False
-
-        # [window setLevel:level] —— 一个 NSInteger 参数
-        msg.argtypes = [ctypes.c_void_p, ctypes.c_void_p, ctypes.c_long]
-        msg(ctypes.c_void_p(window), sel_set_level, level)
-        if level > 0:
-            # Changing WindowStaysOnTopHint recreates the NSWindow. Setting the
-            # floating level alone may leave the replacement ordered behind
-            # the currently active application until Cocoa's next ordering
-            # pass; orderFrontRegardless commits the new level immediately.
-            msg.argtypes = [ctypes.c_void_p, ctypes.c_void_p]
-            msg(ctypes.c_void_p(window), sel_order_front)
-        return True
-    except Exception:
-        return False
 
 
 # 直播捕获兼容模式下窗口标题（普通顶层窗口需要可见标题，供直播姬/OBS 选择）
@@ -348,102 +298,11 @@ def wander_target_y(
     return int(max(y_lo, min(y_hi, start_y + rnd.randint(-max_dy, max_dy))))
 
 
-# ---- Win32：全屏判定用常量/结构 ----
-GWL_STYLE = -16             # GetWindowLongW：取窗口样式
-GWL_EXSTYLE = -20           # GetWindowLongW：取扩展样式
-_WS_CAPTION = 0x00C00000    # WS_BORDER | WS_DLGFRAME（带标题栏）
-_WS_EX_TOPMOST = 0x00000008  # 置顶：真全屏游戏/视频几乎必带，普通最大化窗口不带
-_WS_EX_TRANSPARENT = 0x00000020
-
-
-class _WinRect(ctypes.Structure):
-    _fields_ = [('left', ctypes.c_long), ('top', ctypes.c_long),
-                ('right', ctypes.c_long), ('bottom', ctypes.c_long)]
-
-
-class _WinMonitorInfo(ctypes.Structure):
-    """GetMonitorInfoW 的 MONITORINFO（只读 rcMonitor：显示器完整几何，物理像素）。"""
-    _fields_ = [('cbSize', ctypes.c_ulong), ('rcMonitor', _WinRect),
-                ('rcWork', _WinRect), ('dwFlags', ctypes.c_ulong)]
-
-
-def _set_windows_click_through(hwnd: int, enabled: bool, user32=None) -> bool:
-    """切换 layered HWND 的输入穿透扩展样式。"""
-    user32 = user32 or ctypes.windll.user32
-    style = int(user32.GetWindowLongW(hwnd, GWL_EXSTYLE))
-    updated = style | _WS_EX_TRANSPARENT if enabled else style & ~_WS_EX_TRANSPARENT
-    if updated == style:
-        return False
-    user32.SetWindowLongW(hwnd, GWL_EXSTYLE, updated)
-    return True
-
-
 def _set_speech_bubble_interactive(pet) -> None:
     """按当前是否可打开快速对话，切换气泡鼠标穿透/可点击。"""
     setter = getattr(pet._speech_bubble, "set_interactive", None)
     if callable(setter):
         setter(callable(getattr(pet, "on_open_quick_chat", None)))
-
-
-class WindowsPerPixelInputController:
-    """根据光标所在像素动态切换 layered window 的输入穿透。
-
-    HTTRANSPARENT 只能继续命中当前线程的窗口，无法穿透到其他应用。
-    WS_EX_TRANSPARENT 会让 Windows 在命中时跳过 layered 桌宠窗口；独立
-    定时器在窗口不再收到鼠标消息时仍能检测光标并恢复角色区域交互。
-    """
-
-    NORMAL_POLL_INTERVAL_MS = 10
-    DRAG_POLL_INTERVAL_MS = 100
-
-    def __init__(self, window: "PetWindow") -> None:
-        self._window = window
-        self._timer = QTimer(window)
-        self._timer.setInterval(self.NORMAL_POLL_INTERVAL_MS)
-        self._timer.timeout.connect(self.refresh)
-        self._timer.start()
-
-    def should_click_through(self, global_pos: QPoint) -> bool:
-        win = self._window
-        if win.mouse_through:
-            return True
-        if getattr(win, '_press_global', None) is not None or not win.isVisible():
-            return False
-        local = win.mapFromGlobal(global_pos)
-        if not QRect(0, 0, win.width(), win.height()).contains(local):
-            return False
-        return win._is_transparent_at(local)
-
-    def refresh(self) -> None:
-        try:
-            enabled = self.should_click_through(QCursor.pos())
-            _set_windows_click_through(int(self._window.winId()), enabled)
-        except (AttributeError, OSError, RuntimeError):
-            logging.debug("更新 Windows 逐像素鼠标穿透失败", exc_info=True)
-
-    def set_drag_active(self, active: bool) -> None:
-        """拖拽按下/松手时切换轮询频率。
-
-        拖拽（_press_global 非 None）期间 should_click_through 恒返回 False，
-        每 10ms 轮询纯属空转：降频到 100ms 减少 Win32/QCursor 调用。
-        松手后立即恢复原频率并强制刷新一次穿透状态；非拖拽状态重复调用是 no-op。
-        """
-        if active:
-            if self._timer.interval() != self.DRAG_POLL_INTERVAL_MS:
-                self._timer.setInterval(self.DRAG_POLL_INTERVAL_MS)
-            return
-        if self._timer.interval() == self.NORMAL_POLL_INTERVAL_MS:
-            return
-        self._timer.setInterval(self.NORMAL_POLL_INTERVAL_MS)
-        self.refresh()
-
-    def stop(self) -> None:
-        self._timer.stop()
-        if not self._window.mouse_through:
-            try:
-                _set_windows_click_through(int(self._window.winId()), False)
-            except (AttributeError, OSError, RuntimeError):
-                pass
 
 
 class PetWindow(QWidget):
@@ -1555,122 +1414,40 @@ class PetWindow(QWidget):
             collision_debug.log(runtime_id, 'impulse_apply', pair=message.get('pair', ''),
                                 dv=(dvx, dvy), displacement=(dx, dy), speed=speed)
 
-    _FS_SKIP_CLASSES = {
-        'Progman', 'WorkerW', 'Shell_TrayWnd', 'Shell_SecondaryTrayWnd',
-        'Windows.UI.Core.CoreWindow',  # 开始菜单/通知中心全屏层
-    }
-
     @staticmethod
     def _fullscreen_geometry_hit(l: float, t: float, r: float, b: float,
                                  geom, has_caption: bool, topmost: bool = False) -> bool:
         """覆盖整屏几何，且（无标题栏 或 置顶）= 真全屏。
 
-        判据组合的原因：
-        - 带标题栏的普通/最大化窗口（含 Windows 自动隐藏任务栏场景）不置顶 → 排除；
-        - 真全屏游戏/视频：多数去掉标题栏（无标题栏直接命中）；Unity/UE 系游戏
-          （如绝区零）全屏时保留 WS_CAPTION 样式位但几乎必带 WS_EX_TOPMOST，用
-          置顶位兜住；
-        - 已最大化后按 F11 的窗口（IsZoomed 仍为真、标题栏被清掉）也正常命中。
-
-        geom 兼容 QRect（方法访问）与 win32 RECT（属性访问）。
+        实现已搬至 pet/platform_win.py（批 6-3），此处为兼容性薄委托。
         """
-        if has_caption and not topmost:
-            return False
-        gl = geom.left() if callable(getattr(geom, "left", None)) else geom.left
-        gt = geom.top() if callable(getattr(geom, "top", None)) else geom.top
-        gr = geom.right() if callable(getattr(geom, "right", None)) else geom.right
-        gb = geom.bottom() if callable(getattr(geom, "bottom", None)) else geom.bottom
-        return l <= gl and t <= gt and r >= gr and b >= gb
+        return platform_win._fullscreen_geometry_hit(
+            l, t, r, b, geom, has_caption, topmost)
 
     # ------------------------------------------------------------------
     # 全屏 watcher：后台线程轮询（纯 win32，线程安全）+ 信号回主线程
     # ------------------------------------------------------------------
     def _fg_fullscreen_win32(self) -> bool:
-        """前台窗口是否真全屏。仅返回布尔值，诊断细节见 _fg_fullscreen_probe。"""
-        try:
-            return self._fg_fullscreen_probe()[0]
-        except Exception:
-            return False
+        """前台窗口是否真全屏。仅返回布尔值，诊断细节见 _fg_fullscreen_probe。
+
+        实现已搬至 pet/platform_win.py（批 6-3），此处为兼容性薄委托。
+        """
+        return platform_win._fg_fullscreen_win32()
 
     @staticmethod
     def _fs_user_busy_state() -> tuple[bool, int]:
         """SHQueryUserNotificationState：Windows 自报的全屏/演示忙状态。
 
-        与几何判定互补——几何判定在 DPI 虚拟化、跨屏、DWM 边界差异下可能漏判，
-        而这个 API 是 Windows 自己（Focus Assist/通知静默）判定"用户正在
-        全屏"的依据，游戏和全屏视频都会触发。返回 (是否全屏忙, 原始状态值)。
+        实现已搬至 pet/platform_win.py（批 6-3），此处为兼容性薄委托。
         """
-        if os.name != 'nt':
-            return False, -1
-        try:
-            state = ctypes.c_int(0)
-            hr = ctypes.windll.shell32.SHQueryUserNotificationState(ctypes.byref(state))
-            if hr != 0:  # S_OK
-                return False, -1
-            # 2=QUNS_BUSY(全屏应用运行中) 3=QUNS_RUNNING_D3D_FULL_SCREEN 4=QUNS_PRESENTATION_MODE
-            return state.value in (2, 3, 4), state.value
-        except Exception:
-            return False, -1
+        return platform_win._fs_user_busy_state()
 
     def _fg_fullscreen_probe(self) -> tuple[bool, str]:
         """前台窗口全屏探测，返回 (是否全屏, 诊断描述)。
 
-        可在任意线程调用——不触碰 Qt 对象。判定链：
-        1. foreground_window_info()（vision.py）：排除不可见/最小化/cloaked
-           窗口，取 DWM 框架边界（物理像素，与本进程 DPI awareness 一致）；
-        2. 排除本进程与 shell 窗口；
-        3. 几何判定：窗口覆盖所在显示器完整几何（含任务栏），且无标题栏或置顶；
-        4. 兜底判定：Windows SHQueryUserNotificationState 报告全屏忙状态。
+        实现已搬至 pet/platform_win.py（批 6-3），此处为兼容性薄委托。
         """
-        if os.name != 'nt':
-            return False, "非 Windows"
-        u32 = ctypes.windll.user32
-        # 句柄是 64 位指针：显式声明签名，避免 ctypes 默认 int32 截断
-        u32.MonitorFromWindow.restype = wintypes.HANDLE
-        u32.MonitorFromWindow.argtypes = [wintypes.HWND, wintypes.DWORD]
-        u32.GetWindowLongW.argtypes = [wintypes.HWND, ctypes.c_int]
-        u32.GetClassNameW.argtypes = [wintypes.HWND, ctypes.c_wchar_p, ctypes.c_int]
-        info = vision_mod.foreground_window_info()
-        if not info:
-            return False, "无可判定前台窗口(不可见/最小化/cloaked)"
-        hwnd = info['hwnd']
-        # 排除本进程与其他变体/多开的桌宠进程（置顶小窗，几何不会误判，
-        # 但 SHQueryUserNotificationState 兜底需要进程名兜底排除）
-        proc = info.get('process', '')
-        if info.get('pid') == os.getpid() or proc.lower().startswith('dsh-pet-'):
-            return False, f"前台是桌宠自身 {proc}"
-        # 排除桌面/任务栏等 shell 窗口
-        buf = ctypes.create_unicode_buffer(256)
-        u32.GetClassNameW(hwnd, buf, 256)
-        cls = buf.value
-        if cls in self._FS_SKIP_CLASSES:
-            return False, f"shell 窗口 {cls}"
-
-        style = u32.GetWindowLongW(hwnd, GWL_STYLE)
-        has_caption = bool(style & _WS_CAPTION)
-        exstyle = u32.GetWindowLongW(hwnd, GWL_EXSTYLE)
-        topmost = bool(exstyle & _WS_EX_TOPMOST)
-        x, y, w, h = info['rect']
-        # 窗口所在显示器的完整几何（与 GetWindowRect/DWM 边界同为
-        # 本进程 DPI awareness 下的坐标，天然一致）
-        mon = u32.MonitorFromWindow(hwnd, 2)  # MONITOR_DEFAULTTONEAREST
-        mi = _WinMonitorInfo()
-        mi.cbSize = ctypes.sizeof(_WinMonitorInfo)
-        if not u32.GetMonitorInfoW(mon, ctypes.byref(mi)):
-            return False, f"GetMonitorInfoW 失败 cls={cls}"
-        if self._fullscreen_geometry_hit(
-                x, y, x + w, y + h, mi.rcMonitor, has_caption, topmost):
-            return True, f"几何覆盖 cls={cls} proc={info.get('process', '')}"
-        busy, bstate = self._fs_user_busy_state()
-        if busy:
-            return True, (f"SHQueryUserNotificationState={bstate} "
-                          f"cls={cls} proc={info.get('process', '')}")
-        detail = (f"未命中 cls={cls} proc={info.get('process', '')} "
-                  f"caption={has_caption} topmost={topmost} "
-                  f"rect=({x},{y},{x + w},{y + h}) "
-                  f"monitor=({mi.rcMonitor.left},{mi.rcMonitor.top},"
-                  f"{mi.rcMonitor.right},{mi.rcMonitor.bottom}) busy={bstate}")
-        return False, detail
+        return platform_win._fg_fullscreen_probe()
 
     def _start_fs_watch(self) -> None:
         """启动全屏监视线程（幂等）。"""
